flask_alfredo/flask_alfredo/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
SQLAlchemy==2.0.30
threadpoolctl==3.5.0
typing_extensions==4.11.0
tzdata==2024.1
Werkzeug==3.0.3
